pyproject.toml
[tool.black]
line-length = 240

[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm[tomli]>=6.3"]
build-backend = "setuptools.build_meta"

[project]
name = "synvo_engine"
version = "0.1.0"
authors = [
    { name = "Your Name", email = "youremail@123.com" },
]
description = "A train engine"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "datasets",
    "hf_transfer",
    "transformers",
    "accelerate",
    "pillow",
    "peft",
    "deepspeed",
    "bitsandbytes",
    "triton",
    "loguru",
    "pydantic",
    "jsonlines",
    "librosa",
    "pynvml"
]

[project.optional-dependencies]
pref = [
    "trl @ git+https://github.com/huggingface/trl.git",
]

all = [
    "synvo_engine[pref]",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
exclude = [
    "test"
]

[tool.wheel]
exclude = [
    "test"
]

[project.scripts]
synvo_launch="synvo_engine.launch.cli:main"


[project.urls]
Homepage = "https://xxx.github.io"
Repository = "https://github.com/xxx/yyy"